+++ typescript/widgets/src/walletIntegrations/starknet.ts
@@ -19,7 +19,7 @@ import {
   WarpTypedTransaction,
   chainMetadataToStarknetChain,
 } from '@hyperlane-xyz/sdk';
-import { ProtocolType, assert } from '@hyperlane-xyz/utils';
+import { ProtocolType, assert, sleep } from '@hyperlane-xyz/utils';
 
 import { widgetLogger } from '../logger.js';
 
@@ -55,7 +55,10 @@ export function useStarknetWalletDetails(): WalletDetails {
 
   return useMemo<WalletDetails>(
     () => ({
-      name: connector?.id || 'Starknet Wallet',
+      name:
+        connector?.id === 'argentX'
+          ? 'Ready Wallet'
+          : connector?.name || 'Starknet Wallet',
       logoUrl:
         typeof connector?.icon === 'string'
           ? connector.icon
@@ -113,9 +116,15 @@ export function useStarknetTransactionFns(
   const onSwitchNetwork = useCallback(
     async (chainName: ChainName) => {
       const chainId = multiProvider.getChainMetadata(chainName).chainId;
-      await switchChainAsync({
-        chainId: chainId.toString(),
-      });
+      try {
+        await switchChainAsync({
+          chainId: chainId.toString(),
+        });
+        // Some wallets seem to require a brief pause after switch
+        await sleep(2000);
+      } catch {
+        logger.warn('Failed to switch chain');
+      }
     },
     [multiProvider, switchChainAsync],
   );